`variantType` can give components special styling
Enhancement: if a `variantType` is specified it is
used to generate class names that are prefixed.

This *should* allow multiple different styled versions
of components within a repository.

By default each component should probably have a
`defaultClassName` so that components passing in
children above can give them correct BEM classes.